Add hook for when run is complete
This allows users to run something specific after the run is complete.
In the case of Rails it will automatically drop the newly created
databases. Since the standard `create` and `drop` tasks don't take
parallelization into consideration we must drop the databases when the
run is complete or else they'll be left behind. This isn't a big deal on
CI since new runs will be fresh servers, but locally this can be an
issue / annoyance.